pyproject.toml
[project]
name = "gearcrew-agno"
version = "0.1.0"
description = "AI agent for extracting hiking/backpacking gear information from web sources and storing in GearGraph database"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "agno>=2.3.4",
    "anthropic>=0.75.0",
    "firecrawl-py>=4.9.0",
    "httpx>=0.28.1",
    "langwatch>=0.7.1",
    "langwatch-scenario>=0.7.14",
    "openai>=2.8.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "youtube-transcript-api>=1.2.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["*.test.py", "test_*.py"]
asyncio_mode = "auto"
